Renamed `transactional` into `transaction`. Renamed `message_proof` into `message`.
