+++ crates/fuel-core/src/database/transaction.rs
