+++ crates/fuel-core/src/query/message.rs
