+++ crates/fuel-core/src/query/message/test.rs
